scripts/requirements.txt
absl-py==2.1.0
accelerate==0.29.3
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
boto3==1.34.106
botocore==1.34.106
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
contourpy==1.2.0
cpm-kernels==1.0.11
cycler==0.12.1
datasets==2.15.0
deepspeed==0.12.5
dill==0.3.7
distro==1.9.0
einops==0.5.0
enlighten==1.10.1
entmax==1.3
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.1
fairscale==0.4.13
filelock==3.9.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
func-timeout==4.3.5
fuzzywuzzy==0.18.0
h11==0.14.0
hanziconv==0.3.2
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.19.4
idna==3.4
immutabledict==4.2.0
importlib-resources==6.1.1
importlib_metadata==7.1.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonargparse==3.13.1
jsonlines==4.0.0
kiwisolver==1.4.5
langdetect==1.0.9
langid==1.1.6
lanQ==1.4.2
latexcodec==3.0.0
Levenshtein==0.25.1
lightning-utilities==0.11.0
ltp==4.2.13
ltp-core==0.1.4
ltp-extension==0.1.13
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mmengine-lite==0.10.4
mpmath==1.3.0
mtdata==0.4.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
ninja==1.11.1.1
nltk==3.8
numpy==1.24.1
openai==1.30.1
OpenCC==1.1.7
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
peft==0.7.1
Pillow==9.3.0
platformdirs==4.2.2
portalocker==2.3.0
prefixed==0.7.1
prettytable==3.10.0
proces==0.1.7
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pybtex==0.24.0
pydantic==2.5.2
pydantic_core==2.14.5
pyext==0.7
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.1
pyphen==0.15.0
pypinyin==0.51.0
python-dateutil==2.8.2
python-Levenshtein==0.25.1
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
rapidfuzz==3.9.0
regex==2023.10.3
requests==2.26.0
responses==0.18.0
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rouge-score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sacrebleu==2.4.0
safetensors==0.4.1
scikit-learn==1.2.1
scipy==1.11.4
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sympy==1.12
tabulate==0.9.0
tensor-parallel==2.0.0
termcolor==2.4.0
textstat==0.7.3
threadpoolctl==3.2.0
tiktoken==0.6.0
timeout-decorator==0.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==0.10.3
torchvision==0.16.2+cu118
tqdm==4.64.1
transformers==4.40.0
transformers-stream-generator==0.0.4
triton==2.1.0
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
unbabel-comet==2.2.2
urllib3==1.26.13
wcwidth==0.2.13
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zhon==2.0.2
zipp==3.17.0
